go/rust: Rename subprocess to command
